Parallelizes the processing of batched vs nonbatched vs expensive. The use case is if you have multiple batch filters and multiple nonbatched filters, it shouldn't be waiting for one to finish before starting the other. They should be executed in parallel

I created a field func with 2 batched filters and 2 nonbatched filters to see the impact
Before
```
BenchmarkFilters10Items            30000             67672 ns/op           19236 B/op        300 allocs/op
BenchmarkFilters100Items            5000            277301 ns/op           97238 B/op       1396 allocs/op
BenchmarkFilters1000Items           1000           2410648 ns/op         1026338 B/op      12207 allocs/op
```

After 
```
BenchmarkFilters10Items            30000             52655 ns/op           18434 B/op        291 allocs/op
BenchmarkFilters100Items            5000            246871 ns/op           96384 B/op       1386 allocs/op
BenchmarkFilters1000Items           1000           2124011 ns/op         1026218 B/op      12210 allocs/op
```